requirements.txt
aiomysql==0.1.1
alembic==1.11.1
anyio==3.7.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.2
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
pycparser==2.21
pydantic==1.10.8
PyJWT==1.7.1
PyMySQL==1.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
